% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian.R
\name{julian}
\alias{julian}
\title{Get Julian Day}
\usage{
julian(x = Sys.time())
}
\arguments{
\item{x}{An object of class POSIX or a vector containing POSIX.}
}
\description{
A helper function used to pull the our information from a POSIX object.
}
