% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma.R
\name{sma}
\alias{sma}
\title{Calculate Simple Moving Average}
\usage{
sma(parameter, k, f = mean)
}
\arguments{
\item{parameter}{signal or vector to be smoothed}

\item{k}{the integer number of entries to average over}
}
\description{
Calculate Simple Moving Average
}
\author{
Laura Whitmore
}
