% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeExcel.R
\name{timeExcel}
\alias{timeExcel}
\title{Convert Time Excel}
\usage{
timeExcel(x, tz = "UTC")
}
\description{
Helper function for converting the datetime stamps from Microsoft Excel.
}
\author{
Thomas Bryce Kelly
}
