% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeMatlab.R
\name{timeMatlab}
\alias{timeMatlab}
\title{Convert Time Matlab}
\usage{
timeMatlab(x = Sys.time(), tz = "UTC")
}
\arguments{
\item{x}{The numeric timestamp that will be converted.}
}
\description{
Convert Time Matlab
}
