% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year.R
\name{year}
\alias{year}
\title{Get Year}
\usage{
year(x = Sys.time())
}
\arguments{
\item{x}{An object of class POSIX or a vector containing POSIX.}
}
\description{
A helper function used to pull the our information from a POSIX object.
}
